refactor: remove duplicated type file
