+++ src/logic/contracts/methodIds.ts
@@ -1,5 +1,5 @@
 import { web3ReadOnly as web3 } from 'src/logic/wallets/getWeb3'
-import { DataDecoded, METHOD_TO_ID } from 'src/routes/safe/store/models/types/transactions.d'
+import { DataDecoded, METHOD_TO_ID } from 'src/logic/safe/store/models/types/transactions.d'
 
 export const decodeParamsFromSafeMethod = (data: string): DataDecoded | null => {
   const [methodId, params] = [data.slice(0, 10) as keyof typeof METHOD_TO_ID | string, data.slice(10)]
