+++ src/logic/safe/store/models/types/transaction.ts
@@ -2,7 +2,7 @@ import { List, Map, RecordOf } from 'immutable'
 import { Confirmation } from './confirmation'
 import { GnosisSafe } from 'src/types/contracts/GnosisSafe.d'
 import { DataDecoded, Transfer } from './transactions'
-import { DecodedParams } from 'src/routes/safe/store/models/types/transactions'
+import { DecodedParams } from 'src/logic/safe/store/models/types/transactions'
 
 export enum TransactionTypes {
   INCOMING = 'incoming',
